restructure form fields components
