+++ frontend/src/Components/FormInputField/FormInputField.module.scss
@@ -1,0 +1,21 @@
+@import 'src/constants';
+
+.input_field {
+  padding: 8px;
+  display: block;
+  outline: 1px solid $grey-2;
+  border: none;
+  width: 100%;
+  border-radius: 3px;
+  &:focus {
+    outline: 1px solid $grey-1;
+  }
+}
+
+.errors {
+  outline: 2px solid $red;
+}
+
+.errors_text {
+  color: $red;
+}
