+++ frontend/src/Components/FormInputField/index.ts
@@ -1,0 +1,1 @@
+export { FormInputField } from './FormInputField';
