+++ frontend/src/Components/FormSelect/FormSelect.module.scss
@@ -1,0 +1,22 @@
+@import 'src/constants';
+
+
+.select {
+  padding: 8px;
+  display: block;
+  outline: 1px solid $grey-2;
+  border: none;
+  width: 100%;
+  border-radius: 3px;
+  &:focus {
+    outline: 1px solid $grey-1;
+  }
+}
+
+.error {
+  outline: 2px solid $red;
+}
+
+.error_text {
+  color: $red;
+}
