+++ frontend/src/Components/FormSelect/index.ts
@@ -1,0 +1,1 @@
+export { FormSelect } from './FormSelect';
