+++ frontend/src/Components/FormTextAreaField/FormTextAreaField.tsx
@@ -1,44 +1,45 @@
 import classNames from 'classnames';
-import { FieldValues, UseFormRegister } from 'react-hook-form/dist/types';
 import { Children } from '~/types';
-import styles from './InputField.module.scss';
+import { FieldValues, UseFormRegister } from 'react-hook-form/dist/types';
+import styles from './FormTextAreaField.module.scss';
 
-type types = 'text' | 'number' | 'email' | 'password';
-
-type InputFieldProps = {
+type FormTextAreaFieldProps = {
   children?: Children;
   className?: string;
-  labelClassName?: string;
   inputClassName?: string;
-  type?: types;
+  labelClassName?: string;
+  rows?: number;
+  cols?: number;
   name: string;
   register: UseFormRegister<FieldValues>;
   required?: boolean;
   errors?: Record<string, unknown>;
 };
 
-export function FormInputField({
+export function FormTextAreaField({
   children,
   className,
-  labelClassName,
   inputClassName,
+  labelClassName,
   errors,
+  cols,
   name,
   required,
   register,
-  type = 'text',
-}: InputFieldProps) {
+  rows = 10,
+}: FormTextAreaFieldProps) {
   return (
     <div className={className}>
       <label className={classNames(styles.label, labelClassName)}>
         {children}
-        <input
+        <textarea
           className={classNames(styles.input_field, inputClassName, errors && name in errors && styles.errors)}
-          type={type}
           {...register(name, { required })}
+          rows={rows}
+          cols={cols}
         />
-        {errors && name in errors && <div className={styles.errors_text}>{errors[name].message}</div>}
       </label>
+      {errors && name in errors && <div className={styles.errors_text}>{errors[name].message}</div>}
     </div>
   );
 }
