+++ frontend/src/Components/FormTextAreaField/index.ts
@@ -1,0 +1,1 @@
+export { FormTextAreaField } from './FormTextAreaField';
