+++ frontend/src/Components/InputField/index.ts
@@ -1,2 +1,1 @@
 export { InputField } from './InputField';
-export { FormInputField } from './FormInputField';
