+++ frontend/src/Components/Select/index.ts
@@ -1,2 +1,1 @@
 export { Select } from './Select';
-export { FormSelect } from './FormSelect';
