+++ frontend/src/Components/TextAreaField/index.ts
@@ -1,2 +1,1 @@
 export { TextAreaField } from './TextAreaField';
-export { FormTextAreaField } from './FormTextAreaField';
